audio: ril: add support for asynchronous RIL connecting
This patch enables the audio RIL interface to try to connect to the RIL socket
asynchronously. Althrough this leads to malfunctioning WB-AMR callbacks on the first
few second after boot, the gain by moving the connection process to a new thread is
worth this little risk.

Furthermore, if device maintainers do not agree with me on this risk, they can disable
the seperated thread by defining

	#define RIL_OPEN_SYNCHRONOUS

in their device's samsung_audio.h

Change-Id: I1e57eefcc2344061856ae2fd812c16369439b9b7